+++ examples/chain_client/ibc/client/query/1_ClientState.py
@@ -1,0 +1,21 @@
+import asyncio
+
+from google.protobuf import symbol_database
+
+from pyinjective.async_client import AsyncClient
+from pyinjective.core.network import Network
+
+
+async def main() -> None:
+    network = Network.testnet()
+    client = AsyncClient(network)
+
+    client_id = "07-tendermint-0"
+
+    state = await client.fetch_ibc_client_state(client_id=client_id)
+    print(state)
+
+
+if __name__ == "__main__":
+    symbol_db = symbol_database.Default()
+    asyncio.get_event_loop().run_until_complete(main())